RAG_HuggingFace_FAISS_Project/requirements.txt

transformers
torch
sentence-transformers
faiss-cpu
numpy
